Remove unused files, move webpack config
