--- server/old/middleware/webpack-hmr.js
@@ -1,18 +1,0 @@
-import WebpackHotMiddleware from 'webpack-hot-middleware';
-import applyExpressMiddleware from '../lib/apply-express-middleware';
-import _debug from 'debug';
-
-const debug = _debug('app:server:webpack-hmr');
-
-export default function (compiler, opts) {
-  const middleware = WebpackHotMiddleware(compiler, opts);
-
-  debug('Enable Webpack Hot Module Replacement (HMR).');
-  return function * (next) {
-    let nextStep = yield applyExpressMiddleware(middleware, this.req, this.res);
-
-    if (nextStep && next) {
-      yield* next;
-    }
-  };
-}
